+++ spec/rspec/core/backtrace_formatter_spec.rb
@@ -64,6 +64,38 @@ module RSpec::Core
       end
     end
 
+    describe "#filter_gem" do
+      shared_examples_for "filtering a gem" do |gem_name, path|
+        it 'filters backtrace lines for the named gem' do
+          formatter = BacktraceFormatter.new
+          line      = File.join(path, "lib", "foo.rb:13")
+
+          expect {
+            formatter.filter_gem gem_name
+          }.to change { formatter.exclude?(line) }.from(false).to(true)
+        end
+      end
+
+      context "for a gem installed globally as a system gem" do
+        it_behaves_like "filtering a gem", "foo",
+          "/Users/myron/.gem/ruby/2.1.1/gems/foo-1.6.3.1"
+      end
+
+      context "for a gem installed in a vendored bundler path" do
+        it_behaves_like "filtering a gem", "foo",
+          "/Users/myron/code/my_project/bundle/ruby/2.1.0/gems/foo-0.3.6"
+      end
+
+      context "for a gem installed by bundler as a :git dependency" do
+        it_behaves_like "filtering a gem", "foo",
+          "/Users/myron/code/my_project/bundle/ruby/2.1.0/bundler/gems/foo-2b826653e1f5"
+      end
+
+      context "for a gem sourced from a local path" do
+        it_behaves_like "filtering a gem", "foo", "/Users/myron/code/foo"
+      end
+    end
+
     describe "#format_backtrace" do
       it "excludes lines from rspec libs by default", :unless => RSpec.world.windows_os? do
         backtrace = [